support AT_SYSINFO and AT_SYSINFO_EHDR on all architectures
Not sure why AT_SYSINFO and AT_SYSINFO_EHDR were limited to x86 and
x86/x86_64 respectively.

Linux sets AT_SYSINFO on ia64 and x86 and AT_SYSINFO_EHDR on all architectures.

Removing the the architecture guard on AT_SYSINFO is harmless as the
identifier (32) is not reused on other architectures, so the code will
work correctly without an architecture guard.

Signed-off-by: Patrick Oppenlander <patrick.oppenlander@gmail.com>